Get prices.
This proved to be messier than expected. The short version is that this
is put into the inventory DB as a JSON blob. Two numbers means the price
is a range; otherwise it's a simple price.

More details: The API has a few fields for price, and none of them are
simply "Here's the price to display." Instead, the displayPrice,
listprice, and product_price all are the same "base" price. However, BJs
often sells items for different prices at different stores. The API also
has a list of prices by store ID, along with a bunch of other info,
which greatly bloats the response and slows response time. There is a
condensed field which allows easy mapping store ID to price.
Unfortunately, it requires processing, as it's a semi-colon separated
string in a specific setup, but it's the best option.

Additionally, some items have a price range. This seems to occur when an
item has multiple variants (e.g. sizes). The underlying data is a bit
odd, in that some of these variants don't have the club to price
mapping, forcing us to fall back to displayPrice (in those instances
only).

That the price is ocassionally a range means we must always treat it as
a range. For storage in the database, the min and max price are dumped
as a JSON list into a TEXT column.